Practice 3 Submission
Hi Professor, 

Attached is my submission for Practice 3. Contained in the README.md file is my explanation and answers for the last part of the assignment -- along with this, I recorded the Big O runtime of the algorithm within the comments of Practice03Test.java. Please let me know if you have any questions, and I hope that you have a nice weekend!